# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Secure marketplace for trading and auctioning items in Aethoria.\nCommon and Rare items trade via fixed-price listings. Legendary items trade via auctions.",
        "title": "Market Dragon API",
        "contact": {
            "name": "Market Dragon Team"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/items": {
            "get": {
                "description": "List every item with its current validated Oracle price (when available).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "List items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemListResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Register a new item owned by a guild. The item is not for sale yet. Legendary items are unique (stock is forced to 1).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Register an item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Idempotency key to make retries safe",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Item to register",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/items/{id}": {
            "get": {
                "description": "Get one item, its live price, and its active offer (fixed-price listing or auction) if any.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Get an item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/items/{id}/auction": {
            "post": {
                "description": "Open an auction for a Legendary item. Only one active auction per item is allowed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Open an auction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Idempotency key to make retries safe",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Seller",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OpenAuctionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.AuctionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/items/{id}/bid": {
            "post": {
                "description": "Place a bid on an item's active auction. A new bid must be at least 5% above the current highest bid. Funds are reserved, not debited.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Place a bid",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Idempotency key to make retries safe",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Bid to place",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PlaceBidRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.BidResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/items/{id}/bid/{bid_id}": {
            "delete": {
                "description": "Cancel a bid on an item's active auction. You cannot cancel a bid while you are the highest bidder.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Cancel a bid",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Bid ID",
                        "name": "bid_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Idempotency key to make retries safe",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Bidder",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CancelBidRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/items/{id}/buy": {
            "post": {
                "description": "Buy a fixed-price item. Checks the balance and daily cap, moves money, and transfers the item in one transaction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Buy an item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Idempotency key to make retries safe",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Buyer",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BuyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/items/{id}/list": {
            "post": {
                "description": "Offer a Common/Rare item at a fixed price (a limit order). Legendary items must use an auction instead.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "List an item for sale",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Idempotency key to make retries safe",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Price and seller",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ListForSaleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.ListingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.AuctionResponse": {
            "type": "object",
            "properties": {
                "ends_at": {
                    "type": "string",
                    "example": "2026-01-02T00:00:00Z"
                },
                "highest_bid_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "item_id": {
                    "type": "integer",
                    "example": 3
                },
                "seller_guild_id": {
                    "type": "integer",
                    "example": 3
                },
                "starts_at": {
                    "type": "string",
                    "example": "2026-01-01T00:00:00Z"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "winner_guild_id": {
                    "type": "integer"
                }
            }
        },
        "dto.BidResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 1200
                },
                "auction_id": {
                    "type": "integer",
                    "example": 1
                },
                "bidder_guild_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
        "dto.BuyRequest": {
            "type": "object",
            "required": [
                "buyer_guild_id"
            ],
            "properties": {
                "buyer_guild_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.CancelBidRequest": {
            "type": "object",
            "required": [
                "bidder_guild_id"
            ],
            "properties": {
                "bidder_guild_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.CreateItemRequest": {
            "type": "object",
            "required": [
                "name",
                "owner_guild_id",
                "tier"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Elven Bow"
                },
                "owner_guild_id": {
                    "type": "integer",
                    "example": 2
                },
                "stock": {
                    "type": "integer",
                    "example": 5
                },
                "tier": {
                    "type": "string",
                    "example": "rare"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "item not found"
                }
            }
        },
...
        "dto.ItemDetailResponse": {
            "type": "object",
            "properties": {
                "auction": {
                    "$ref": "#/definitions/dto.AuctionResponse"
                },
                "item": {
                    "$ref": "#/definitions/dto.ItemResponse"
                },
                "listing": {
                    "$ref": "#/definitions/dto.ListingResponse"
                }
            }
        },
        "dto.ItemListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ItemResponse"
                    }
                }
            }
        },
        "dto.ItemResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 3
                },
                "name": {
                    "type": "string",
                    "example": "Soul Reaver"
                },
                "owner_guild_id": {
                    "type": "integer",
                    "example": 3
                },
                "price": {
                    "type": "integer",
                    "example": 250000
                },
                "status": {
                    "type": "string",
                    "example": "available"
                },
                "stock": {
                    "type": "integer",
                    "example": 1
                },
                "tier": {
                    "type": "string",
                    "example": "legendary"
                }
            }
        },
        "dto.ListForSaleRequest": {
            "type": "object",
            "required": [
                "price",
                "seller_guild_id"
            ],
            "properties": {
                "price": {
                    "type": "integer",
                    "example": 500
                },
                "seller_guild_id": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "dto.ListingResponse": {
            "type": "object",
            "properties": {
                "buyer_guild_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "item_id": {
                    "type": "integer",
                    "example": 2
                },
                "price": {
                    "type": "integer",
                    "example": 500
                },
                "seller_guild_id": {
                    "type": "integer",
                    "example": 2
                },
                "status": {
                    "type": "string",
                    "example": "open"
                }
            }
        },
        "dto.OpenAuctionRequest": {
            "type": "object",
            "required": [
                "seller_guild_id"
            ],
            "properties": {
                "seller_guild_id": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "dto.PlaceBidRequest": {
            "type": "object",
            "required": [
                "amount",
                "bidder_guild_id"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 1200
                },
                "bidder_guild_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "cancelled"
                }
            }
        },
...
    },
...
}